requirements.txt
einops==0.8.0
hydra-core==1.3.2
keras==2.10.0
matplotlib==3.10.0
numpy==2.2.1
omegaconf==2.3.0
pysindy==1.7.5
scipy==1.15.0
tensorflow==2.10.1
